Move all argparsing logic to separate file
